Moving SaveSystem to separate project
