+++ CMakeLists.txt
@@ -102,6 +102,7 @@ add_subdirectory(OpenGLRendering)
 add_subdirectory(AudioEngine)
 add_subdirectory(CSC8508)
 add_subdirectory(ImGui)
+add_subdirectory(SerializedSave)
 add_subdirectory(UISystem)
 if(USE_VULKAN)
     add_subdirectory(VulkanRendering)
