+++ CSC8508/CMakeLists.txt
@@ -178,6 +178,7 @@ include_directories("../CSC8508CoreClasses/")
 include_directories("../Event/")
 include_directories("../ImGui/")
 include_directories("../UISystem/")
+include_directories("../SerializedSave/")
 
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC NCLCoreClasses)
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC CSC8508CoreClasses)
@@ -187,6 +188,7 @@ target_link_libraries(${PROJECT_NAME} LINK_PUBLIC Event)
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC AudioEngine)
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC ImGui)
 target_link_libraries(${PROJECT_NAME} LINK_PUBLIC UISystem)
+target_link_libraries(${PROJECT_NAME} LINK_PUBLIC SerializedSave)
 
 # Handle `fmod.dll` copying here instead of `AudioEngine`
 add_custom_command(TARGET CSC8508 POST_BUILD
