+++ CSC8508CoreClasses/GameObject.h
@@ -6,7 +6,7 @@
 #include "Transform.h"
 #include "ComponentManager.h"
 #include <vector>
-#include "ISerializable.h"
+#include "../SerializedSave/ISerializable.h"
 
 using std::vector;
