+++ CSC8508CoreClasses/IComponent.h
@@ -7,7 +7,6 @@
 
 #include "Transform.h"
 #include "GameObject.h"
-//#include "ISerializable.h"
 #include <unordered_set>
 
 namespace NCL::CSC8508
