+++ NCLCoreClasses/CMakeLists.txt
@@ -123,8 +123,8 @@ target_precompile_headers(${PROJECT_NAME} PRIVATE
     <thread>
     <filesystem>
     <functional>
-	<algorithm>
-	<assert.h>  
+    <algorithm>
+    <assert.h>  
 )
 
 ################################################################################
