+++ SerializedSave/CMakeLists.txt
@@ -1,0 +1,6 @@
+﻿file(GLOB Headers "${CMAKE_CURRENT_SOURCE_DIR}/*.h")
+file(GLOB Sources "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
+
+add_library(SerializedSave
+${Headers}
+${Sources})
