+++ SerializedSave/SaveManager.cpp
@@ -1,0 +1,1 @@
+#include "SaveManager.h"
